更改 Group ID 属性

+++ cmd/bot/internal/auth/auth.go
@@ -4,7 +4,7 @@ func IsCreator (creator int,uid int) bool {
 	return uid == creator
 }
 
-func (cfg Config) IsAuthGroups (gid int) bool {
+func (cfg Config) IsAuthGroups (gid int64) bool {
 	for _,authGid := range cfg.Groups {
 		return gid == authGid
 	}
